+++ online-store/src/components/Filter.ts
@@ -1,9 +1,9 @@
 /* eslint-disable no-param-reassign */
 import sliders from './sliders';
 import Catalog from './Catalog';
-import books from './books.json';
-import { FilterType, FilterNames } from './types';
-import { minFrom, maxFrom, createCard } from './utils';
+import books from '../data/books.json';
+import { FilterType, FilterNames } from '../helpers/types';
+import { minFrom, maxFrom, createCard } from '../helpers/utils';
 
 export default class Filter {
   // eslint-disable-next-line no-use-before-define
