+++ online-store/src/index.ts
@@ -1,8 +1,8 @@
 import './style.css';
 import Catalog from './components/Catalog';
 import Filter from './components/Filter';
-import { FilterNames } from './components/types';
-import { convertToNumbers } from './components/utils';
+import { FilterNames } from './helpers/types';
+import { convertToNumbers } from './helpers/utils';
 import sliders from './components/sliders';
 import Cart from './components/Cart';
 import Resetter from './components/Resetter';
